pyproject.toml
[project]
name = "cairn"
version = "0.1.0"
description = "Personal autonomous AI agent with Shared Context Memory Store"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.1",
    "croniter>=6.2.2",
    "docker>=7.1.0",
    "duckduckgo-search>=7.0.0",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "supabase>=2.28.2",
    "trafilatura>=2.0.0",
    "fastmcp>=2.0.0",
]

[project.scripts]
cairn = "main:main"
